requirements.txt
# image classifier
Pillow

# attributes scraper
aiohttp
nordvpn-switcher

# pretty stuff
rich
